+++ package.json
@@ -11,8 +11,10 @@
     "react-icons": "^5.3.0",
     "react-router-dom": "^6.26.2",
     "react-scripts": "5.0.1",
+    "react-social-login-buttons": "^4.1.0",
     "styled-components": "^6.1.13",
-    "web-vitals": "^2.1.0"
+    "web-vitals": "^2.1.0",
+    "zustand": "^5.0.0-rc.2"
   },
   "scripts": {
     "start": "react-scripts start",
