+++ src/routes/AppRoutes.js
@@ -5,12 +5,13 @@ import Login from '../pages/Login';
 import MainLayout from '../layouts/MainLayout';
 import AuthLayout from '../layouts/AuthLayout';
 import Maintenance from '../components/MaintenancePage';
+
 const AppRoutes = () => {
   return (
     <Routes> {/* Switch 대신 Routes 사용 */}
       {/* 인증 레이아웃이 필요한 경우 */}
-      <Route path="/" element={<MainLayout><Maintenance /></MainLayout>} />
-      <Route path="/login" element={<AuthLayout><Login /></AuthLayout>} />
+      <Route path="/" element={<MainLayout><Maintenance/></MainLayout>} />
+      <Route path="/login" element={<AuthLayout><Login /></AuthLayout>} /> 
       
       {/* 다른 페이지 경로 추가 */}
     </Routes>
